Rm repeated code
